Floresta and Penn treebank datasets
Here we are using Floresta dataset which is a Portuguese treebank containing manually annotated sentences with their corresponding POS tags.  The Penn Treebank (PTB) is one of the most widely used datasets in Natural Language Processing (NLP) and Computational Linguistics. It consists of a large corpus of English text that has been meticulously annotated with syntactic and structural information, including Part-of-Speech (POS) tags, phrase structure trees, and syntactic dependencies.